+++ src/module/settings.ts
@@ -6,6 +6,7 @@ export const VARIANT_ENCUMBRANCE_FLAG = 'variant-encumbrance-dnd5e';
 export const VARIANT_ENCUMBRANCE_INVENTORY_PLUS_MODULE_NAME = 'inventory-plus';
 export const VARIANT_ENCUMBRANCE_MIDI_QOL_MODULE_NAME = 'midi-qol';
 export const VARIANT_ENCUMBRANCE_ITEM_COLLECTION_MODULE_NAME = 'itemcollection';
+export const VARIANT_ENCUMBRANCE_DFREDS_CONVENIENT_EFFECTS_MODULE_NAME = 'dfreds-convenient-effects';
 
 /**
  * Because typescript doesn't know when in the lifecycle of foundry your code runs, we have to assume that the